Added Player Movement
I added player movement to the game. The player is restricted to the grid (not tested), but not to the available platforms. That's what should be added next.

I don't think the player will be stored in the grid. Everything else will be, but the player will not be in any array, though the player object will save the player's index in the game grid.