Floor Origin is Now the Centre of the Map
The floor origin is now the map centre instead of the top left of the map. I may want to zoom in the camera based on how big the floor is, but I'll leave it as is for now.

The camera is currently at a bird's-eye view. In the final game, it will be at an isometric view, so that will be tested out later. I might also change the total amount of blocks in the world later.