Tested Attacks and Enemies, Made Block Prefabs
I tested out the player's attack function, and it works. I also tested the enemy implementation, and both the stationary enemy and the copy enemy works. However, the enemy proper IDs will need to be figured out later.

I changed the blocks to be prefab variants so that edits can be shared between them.

I set up hitboxes for the player, enemies, and blocks. The player and enemy hitboxes are meant for the enemy to kill the player, but the block hitboxes don't do anything. I might change collisions in the game so that they all use 3D hitboxes instead of manually calling the interact function so that it's more consistent.