Changed Switch Block,
The switch block has been changed so that it switches when the player attacks, not when they jump. The toggle works, and it kills entities that are on top of it properly.

I made some slight quality of use changes to the limited block, and I now have the phase block get turned active/inactive until I add an animation for it.