Added Skybox Feature
The ability to change the skybox has been added to the game. There's only a debug skybox for now for testing purposes, and I don't know if I'll have more than one skybox for the gameplay scene, but the option is there. I've also added a slight outline for the text.