Added Weapon Item + Player Reset Anim Fix + Working On Copy Anims
I added the weapon item so that the player can start off being unable to attack on a floor.

I fixed a bug where the player's idle animation wasn't being used when an entity was reset.

I'm working on fixing the copy enemy's animations since the jump animation plays even when the copy enemy doesn't jump.