Worked on the Game Settings
I added in functionality for changing the BGM and SFX volume, as well as muting/unmuting the game. I may have the ability to export saved settings to a file, but I'm not sure (it isn't really necessary).

I should get some base button assets before I start adding more so that I don't have to replace that many with custom designs later on.