Set Up Boss Goal + Boss Dies On Contact for No Reason
The goal has been adjusted so that it properly gets enabled when the boss dies. However, the boss dies on contact with the player for some reason. I don't know why this happens, but it does.

It gets destroyed when the floor fail function is called, and gets recreated on reset, but then it just gets deleted. The OnDestroy function doesn't appear to be called, but it doesn't show up in the list, so the object should be gone. I need to fix that.