Fixed Final Boss Deletion Glitch
Due to the way the game is set up, the final boss can't be an instance, otherwise it gets deleted. This is because the old one is not removed from memory before the new one is made, causing both of them to get deleted since the program thinks one already exists. This has been fixed.