Added Floor Codes + Floor Count Changer
I added the rest of the floor codes, so now each floor can be selected. However, not all floors have any geometry yet.

I've also added a variable to check against the floor count so that I can limit how many floor clears are needed to beat the game. However, this won't be used in the final game. The player will have to clear all 50 floors of the game to beat it.