Working on Bar Enemy
Made a material with the emissive value for the bar enemy segment. Relatedly, I have encountered a problem where the bar enemy segment will still trigger collision even if the player is in the process of moving away from it. This is due to the bar enemy's movement being instant. This will need to be fixed.