Added a Tutorial Toggle + Adjusted Save System + Adjusted Pause Fixes
Added a tutorial toggle to the settings menu. The save system has also been adjusted so that it remembers if the game is complete or not. If the game is complete, a continued game starts from the beginning.

I have also adjusted the pause fixes so that the options window doesn't disappear if the player decides to go to the title screen or the results screen. The pause window is off by default, so the unpause call in post start was left as is.