Added Skyboxes
The images have been put together to add the rest of the skyboxes. The seams on the skyboxes seem to be screwed up a bit, so I'll need to see if I can fix that, but the skyboxes are in now.

Honestly, they don't look as good as I hoped, but there's nothing I can do about that.