Fixed Tutorial Timescale Problem + Tutorial Diagram Problem
Fixed a problem where the tutorial timescale was not being kept at 0 when the tutorial started. This was because of a conflict with the floor loading coroutine.

Fixed an issue where the tutorial diagram was not showing up. This was done by removing the diagram changes when the text box opens and closes, instead regulating it to the pages.