Partial Fix to SFXs at Floor Start
A partial fix has been implemented for sound effects that play at the start of a floor. It seems to work for everything but the goal being locked, which is briefly visible when the game begins. Maybe try adding back in the yield call in the Floor Coroutine and seeing if that fixes it.

I also fixed a bug where the floor loading screen wasn't being shown on the first floor of a continued game.